+++ cli/build.clj
@@ -2,10 +2,10 @@
   (:require
    [clojure.java.io :as io]
    [clojure.string :as string]
-   [clojure.tools.build.api :as b]
-   [deps-bin.impl.bin :as deps-bin]))
+   [clojure.tools.build.api :as b]))
 
 (def lib 'com.github.clojure-lsp/clojure-lsp)
+(def clojars-lib 'com.github.clojure-lsp/clojure-lsp-standalone)
 (def current-version (string/trim (slurp (io/resource "CLOJURE_LSP_VERSION"))))
 (def class-dir "target/classes")
 (def basis {:project "deps.edn"})
@@ -24,6 +24,23 @@
 (defn ^:private uber [opts]
   (clean opts)
   (javac opts)
+  (let [basis (b/create-basis (update basis :aliases concat (:extra-aliases opts)))]
+    (b/write-pom {:target ""
+                  :lib clojars-lib
+                  :version current-version
+                  :basis basis
+                  :src-dirs ["src" "../lib/src"]
+                  :resource-dirs ["resources"]})
+    (b/copy-dir {:src-dirs ["src" "../lib/src" "resources" "../lib/resources"]
+                 :target-dir class-dir})
+    (b/uber {:class-dir class-dir
+             :uber-file uber-file
+             :main 'clojure-lsp.main
+             :basis basis})))
+
+(defn ^:private uber-aot [opts]
+  (clean opts)
+  (javac opts)
   (println "Building uberjar...")
   (let [basis (b/create-basis (update basis :aliases concat (:extra-aliases opts)))]
     (b/copy-dir {:src-dirs ["src" "resources"]
@@ -39,18 +56,19 @@
 
 (defn ^:private bin []
   (println "Generating bin...")
-  (deps-bin/build-bin {:jar uber-file
-                       :name "clojure-lsp"
-                       :jvm-opts ["-Xmx2g" "-server"]
-                       :skip-realign true}))
+  ((requiring-resolve 'deps-bin.impl.bin/build-bin)
+   {:jar uber-file
+    :name "clojure-lsp"
+    :jvm-opts ["-Xmx2g" "-server"]
+    :skip-realign true}))
 
-(def prod-jar uber)
+(def prod-jar uber-aot)
 
 (defn prod-jar-for-native [opts]
-  (uber (merge opts {:extra-aliases [:native]})))
+  (uber-aot (merge opts {:extra-aliases [:native]})))
 
 (defn debug-cli [opts]
-  (uber (merge opts {:extra-aliases [:debug]}))
+  (uber-aot (merge opts {:extra-aliases [:debug]}))
   (bin))
 
 (defn prod-cli [opts]
@@ -80,3 +98,12 @@
       (when-not (= 0 exit)
         (System/exit exit)))
     (println "Set GRAALVM_HOME env")))
+
+(defn deploy-clojars [opts]
+  (uber opts)
+  ((requiring-resolve 'deps-deploy.deps-deploy/deploy)
+   (merge {:installer :remote
+           :artifact uber-file
+           :pom-file "pom.xml"}
+          opts))
+  opts)