+++ lib/test/clojure_lsp/feature/move_coll_entry_test.clj
@@ -2,33 +2,36 @@
   (:require
    [clojure-lsp.db :as db]
    [clojure-lsp.feature.move-coll-entry :as f.move-coll-entry]
-   [clojure-lsp.parser :as parser]
    [clojure-lsp.test-helper :as h]
    [clojure.test :refer [deftest is testing]]
-   [rewrite-clj.node :as n]
    [rewrite-clj.zip :as z]))
 
 (h/reset-db-after-test)
 
 (def ^:private uri (h/file-uri "file:///a.cljc"))
 
-(defn zloc-at [row col]
+(defn zloc-at [pos]
   (-> @db/db
       (get-in [:documents uri])
       :text
-      (parser/loc-at-pos row col)))
+      (z/of-string {:track-position? true})
+      (z/find-last-by-pos pos)))
 
 (defn load-code-and-zloc [code]
-  (let [[[row col]] (h/load-code-and-locs code uri)]
-    (zloc-at row col)))
+  (let [[pos] (h/load-code-and-locs code uri)]
+    (zloc-at pos)))
+
+(defn can-move-zloc-up? [zloc]
+  (f.move-coll-entry/can-move-entry-up? zloc uri @db/db))
+
+(defn can-move-zloc-down? [zloc]
+  (f.move-coll-entry/can-move-entry-down? zloc uri @db/db))
 
 (defn can-move-code-up? [code]
-  (let [zloc (load-code-and-zloc code)]
-    (f.move-coll-entry/can-move-entry-up? zloc uri @db/db)))
+  (can-move-zloc-up? (load-code-and-zloc code)))
 
 (defn can-move-code-down? [code]
-  (let [zloc (load-code-and-zloc code)]
-    (f.move-coll-entry/can-move-entry-down? zloc uri @db/db)))
+  (can-move-zloc-down? (load-code-and-zloc code)))
 
 (deftest can-move-entry-up?
   (testing "common cases"
@@ -235,32 +238,34 @@
                             "      :let [{:keys [b]} {:b 1}"
                             "            |{:keys [c]} {:c 1}]])")))
   (testing "blank lines"
-    (is (= (h/code "{:b 2"
-                   ""
-                   " :a 1}")
-           (-> (h/code "{:a 1"
-                       "|"
-                       " :b 2}")
-               load-code-and-zloc
-               ;; load-code moves cursor in whitespace to outer form;
-               ;; move cursor to blank line between a and b
-               (z/down)
-               (z/find-next-value z/right 1)
-               (z/right*)
-               move-zloc-up
-               as-string)))
-    (is (nil? (-> (h/code "{:a 1"
-                          ""
-                          " :b 2"
-                          "|"
-                          "}")
-                  load-code-and-zloc
-                  ;; load-code moves cursor in whitespace to outer form;
-                  ;; move cursor to blank line after b
-                  (z/down)
-                  (z/find-next-value z/right 2)
-                  z/right*
-                  move-zloc-up))))
+    (assert-move-up (h/code "{:b 2"
+                            ""
+                            " :a 1}")
+                    (h/code "{:a 1"
+                            "|"
+                            " :b 2}"))
+    (assert-move-up (h/code "{:a 1"
+                            ""
+                            " :c 2"
+                            ""
+                            " :b 2}")
+                    (h/code "{:a 1"
+                            ""
+                            " :b 2"
+                            "|"
+                            " :c 2}"))
+    ;; NOTE: ideally can-move-*? and move-* would always agree, but when they
+    ;; don't at least no erroneous swaps happen
+    (let [ws-zloc (load-code-and-zloc (h/code "[:a"
+                                              "|"
+                                              "]"))]
+      (is (can-move-zloc-up? ws-zloc))
+      (is (nil? (move-zloc-up ws-zloc))))
+    (let [ws-zloc (load-code-and-zloc (h/code "{:a 1"
+                                              "|"
+                                              "}"))]
+      (is (can-move-zloc-up? ws-zloc))
+      (is (nil? (move-zloc-up ws-zloc)))))
   (testing "comments"
     (assert-move-up (h/code "{:b (+ 1 1) ;; two comment"
                             " :a 1 ;; one comment"
@@ -331,46 +336,43 @@
                             " }")
                     (h/code "{:a 1 ;; one comment"
                             " |:b 2}"))
-    ;; moves from leading comment / whitespace
-    (is (= (h/code "{;; b comment"
-                   " :b 2 ;; two comment"
-                   " :a 1 ;; one comment"
-                   " :c 3}")
-           (-> (h/code "{:a 1 ;; one comment"
-                       " ;; |b comment"
-                       " :b 2 ;; two comment"
-                       " :c 3}")
-               load-code-and-zloc
-               ;; load-code moves cursor in whitespace to outer form;
-               ;; move cursor to <comment '; b comment'>
-               (z/down)
-               (z/find-next-value z/right :b)
-               (z/find z/left* (comp n/comment? z/node))
-               move-zloc-up
-               as-string)))
-    ;; moves from trailing comment / whitespace
-    (is (= (h/code "{:b 2 ;; two comment"
-                   " :a 1 ;; one comment"
-                   " :c 3}")
-           (-> (h/code "{:a 1 ;; one comment"
-                       " :b 2 ;; |two comment"
-                       " :c 3}")
-               load-code-and-zloc
-               ;; load-code moves cursor in whitespace to outer form;
-               ;; move cursor to <comment '; two comment'>
-               (z/down)
-               (z/find-next-value z/right :b)
-               (z/find z/right* (comp n/comment? z/node))
-               move-zloc-up
-               as-string))))
+    ;; moves from leading comment
+    (assert-move-up (h/code "{;; b comment"
+                            " :b 2 ;; two comment"
+                            " :a 1 ;; one comment"
+                            " :c 3}")
+                    (h/code "{:a 1 ;; one comment"
+                            " ;; |b comment"
+                            " :b 2 ;; two comment"
+                            " :c 3}"))
+    ;; moves from trailing comment
+    (assert-move-up (h/code "{:b 2 ;; two comment"
+                            " :a 1 ;; one comment"
+                            " :c 3}")
+                    (h/code "{:a 1 ;; one comment"
+                            " :b 2 ;; |two comment"
+                            " :c 3}"))
+    ;; NOTE: ideally can-move-*? and move-* would always agree, but when they
+    ;; don't at least no erroneous swaps happen
+    (let [ws-zloc (load-code-and-zloc (h/code "[:a |;; a comment"
+                                              "]"))]
+      (is (can-move-zloc-up? ws-zloc))
+      (is (nil? (move-zloc-up ws-zloc))))
+    (let [ws-zloc (load-code-and-zloc (h/code "{:a 1 |;; one comment"
+                                              "}"))]
+      (is (can-move-zloc-up? ws-zloc))
+      (is (nil? (move-zloc-up ws-zloc)))))
   (testing "relocation"
     (assert-move-up-position [1 2]
-                             (h/code "{:a 1 ;; one comment"
+                             (h/code "{:a 1"
                                      " |:b 2}"))
+    (assert-move-up-position [1 2]
+                             (h/code "[:a"
+                                     " |:b]"))
     ;; moves cursor to start of entry pair
     (assert-move-up-position [1 2]
-                             (h/code "{:a :x ;; one comment"
-                                     " :b    |:y}"))))
+                             (h/code "{:a 1"
+                                     " :b |2}"))))
 
 (deftest move-down
   (testing "common cases"
@@ -468,31 +470,34 @@
                               "      :let [|{:keys [b]} {:b 1}"
                               "            {:keys [c]} {:c 1}]])")))
   (testing "blank lines"
-    (is (= (h/code "{"
-                   " :b 2"
-                   ""
-                   " :a 1}")
-           (-> (h/code "{|"
-                       " :a 1"
-                       ""
-                       " :b 2}")
-               load-code-and-zloc
-               ;; load-code moves cursor in whitespace to outer form;
-               ;; move cursor to blank line above a
-               (z/down*)
-               move-zloc-down
-               as-string)))
-    (is (nil? (-> (h/code "{:a 1"
-                          "|"
-                          " :b 2}")
-                  load-code-and-zloc
-                  ;; load-code moves cursor in whitespace to outer form;
-                  ;; move cursor to blank line between a and b
-                  z/down
-                  z/right
-                  z/right*
-                  z/right*
-                  move-zloc-down))))
+    (assert-move-down (h/code "{"
+                              " :b 2"
+                              ""
+                              " :a 1}")
+                      (h/code "{|"
+                              " :a 1"
+                              ""
+                              " :b 2}"))
+    (assert-move-down (h/code "{:a 1"
+                              ""
+                              " :c 2"
+                              ""
+                              " :b 2}")
+                      (h/code "{:a 1"
+                              "|"
+                              " :b 2"
+                              ""
+                              " :c 2}"))
+    ;; NOTE: ideally can-move-*? and move-* would always agree, but when they
+    ;; don't at least no erroneous swaps happen
+    (let [ws-zloc (load-code-and-zloc (h/code "[|"
+                                              " :a]"))]
+      (is (can-move-zloc-down? ws-zloc))
+      (is (nil? (move-zloc-down ws-zloc))))
+    (let [ws-zloc (load-code-and-zloc (h/code "{|"
+                                              " :a 1}"))]
+      (is (can-move-zloc-down? ws-zloc))
+      (is (nil? (move-zloc-down ws-zloc)))))
   (testing "comments"
     (assert-move-down (h/code "{:b (+ 1 1) ;; two comment"
                               " :a 1 ;; one comment"
@@ -563,42 +568,42 @@
                               " }")
                       (h/code "{|:a 1 ;; one comment"
                               " :b 2}"))
-    ;; moves from leading comment / whitespace
-    (is (= (h/code "{;; b comment"
-                   " :b 2 ;; two comment"
-                   " ;; a comment"
-                   " :a 1 ;; one comment"
-                   " :c 3}")
-           (-> (h/code "{;; |a comment"
-                       " :a 1 ;; one comment"
-                       " ;; b comment"
-                       " :b 2 ;; two comment"
-                       " :c 3}")
-               load-code-and-zloc
-               ;; load-code moves cursor in whitespace to outer form;
-               ;; move cursor to <comment '; a comment'>
-               (z/down*)
-               move-zloc-down
-               as-string)))
-    ;; moves from trailing comment / whitespace
-    (is (= (h/code "{:b 2 ;; two comment"
-                   " :a 1 ;; one comment"
-                   " :c 3}")
-           (-> (h/code "{:a 1 ;; |one comment"
-                       " :b 2 ;; two comment"
-                       " :c 3}")
-               load-code-and-zloc
-               ;; load-code moves cursor in whitespace to outer form;
-               ;; move cursor to <comment '; one comment'>
-               (z/down)
-               (z/find z/right* (comp n/comment? z/node))
-               move-zloc-down
-               as-string))))
+    ;; moves from leading comment
+    (assert-move-down (h/code "{;; b comment"
+                              " :b 2 ;; two comment"
+                              " ;; a comment"
+                              " :a 1 ;; one comment"
+                              " :c 3}")
+                      (h/code "{;; |a comment"
+                              " :a 1 ;; one comment"
+                              " ;; b comment"
+                              " :b 2 ;; two comment"
+                              " :c 3}"))
+    ;; moves from trailing comment
+    (assert-move-down (h/code "{:b 2 ;; two comment"
+                              " :a 1 ;; one comment"
+                              " :c 3}")
+                      (h/code "{:a 1 ;; |one comment"
+                              " :b 2 ;; two comment"
+                              " :c 3}"))
+    ;; NOTE: ideally can-move-*? and move-* would always agree, but when they
+    ;; don't at least no erroneous swaps happen
+    (let [ws-zloc (load-code-and-zloc (h/code "[|;; a comment"
+                                              " :a]"))]
+      (is (can-move-zloc-down? ws-zloc))
+      (is (nil? (move-zloc-down ws-zloc))))
+    (let [ws-zloc (load-code-and-zloc (h/code "{|;; a comment"
+                                              " :a 1}"))]
+      (is (can-move-zloc-down? ws-zloc))
+      (is (nil? (move-zloc-down ws-zloc)))))
   (testing "relocation"
     (assert-move-down-position [2 2]
-                               (h/code "{|:a 1 ;; one comment"
+                               (h/code "{|:a 1"
                                        " :b 2}"))
+    (assert-move-down-position [2 2]
+                               (h/code "[|:a"
+                                       " :b]"))
     ;; moves cursor to start of entry pair
     (assert-move-down-position [2 2]
-                               (h/code "{:a    |:x ;; one comment"
-                                       " :b :y}"))))
+                               (h/code "{:a |1"
+                                       " :b 2}"))))